requirements.txt
pytest
pytest-selenium

